+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## v2.2.1 - 2019-02-20
 ### Fixed
 * Fix the display of archived group links on the student dashboard ([#907](https://gitlab.com/yale-sdmp/vesta/issues/907)).
+* Fix the drawless suites query ([#901](https://gitlab.com/yale-sdmp/vesta/issues/901)).
 
 ## v2.2.0 - 2019-02-12
 ### Added